requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.8.0
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
coqpit-config==0.2.0
coqui-tts
coqui-tts-trainer==0.2.2
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
dateparser==1.1.8
dearpygui==2.0.0
decorator==5.1.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.17.0
fonttools==4.55.5
frozenlist==1.5.0
fsspec==2024.12.0
googletrans==4.0.0rc1
grpcio==1.69.0
gruut
gruut-ipa
gruut_lang_de
gruut_lang_en
gruut_lang_es
gruut_lang_fr
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.27.1
hyperframe==5.2.0
idna==2.10
inflect==7.5.0
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monotonic-alignment-search==0.1.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
shellingham==1.5.4
silero==0.4.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.5.0.post1
speakerpy==0.1.2
SpeechRecognition==3.13.0
srsly==2.5.1
SudachiDict-core==20241021
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio
tqdm==4.67.1
transformers==4.46.2
transliterate==1.10.2
triton
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
wasabi==1.1.3
Wave==0.0.2
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
